Moved Secret to kernel module
