+++ delta/rdf/src/main/scala/ch/epfl/bluebrain/nexus/delta/rdf/implicits/package.scala
@@ -1,6 +1,6 @@
 package ch.epfl.bluebrain.nexus.delta.rdf
 
-import ch.epfl.bluebrain.nexus.delta.rdf.instances.{TripleInstances, UriInstances}
+import ch.epfl.bluebrain.nexus.delta.rdf.instances.{SecretInstances, TripleInstances, UriInstances}
 import ch.epfl.bluebrain.nexus.delta.rdf.syntax.{IriSyntax, IterableSyntax, JsonLdEncoderSyntax, JsonSyntax, PathSyntax, UriSyntax}
 
 package object implicits
@@ -12,3 +12,4 @@ package object implicits
     with IterableSyntax
     with UriSyntax
     with PathSyntax
+    with SecretInstances
