+++ delta/rdf/src/main/scala/ch/epfl/bluebrain/nexus/delta/rdf/instances/SecretInstances.scala
@@ -1,0 +1,9 @@
+package ch.epfl.bluebrain.nexus.delta.rdf.instances
+
+import ch.epfl.bluebrain.nexus.delta.kernel.Secret
+import ch.epfl.bluebrain.nexus.delta.rdf.jsonld.decoder.JsonLdDecoder
+
+trait SecretInstances {
+  implicit def secretJsonLdDecoder[A](implicit D: JsonLdDecoder[A]): JsonLdDecoder[Secret[A]] =
+    D.map(Secret.apply)
+}
