+++ delta/rdf/src/main/scala/ch/epfl/bluebrain/nexus/delta/rdf/instances/package.scala
@@ -1,3 +1,3 @@
 package ch.epfl.bluebrain.nexus.delta.rdf
 
-package object instances extends TripleInstances with UriInstances
+package object instances extends TripleInstances with UriInstances with SecretInstances
